+++ shared/domain/seasons/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/seasons/api/model/SeasonUiModel.kt
@@ -1,8 +1,8 @@
 package com.thomaskioko.tvmaniac.seasons.api.model
 
 data class SeasonUiModel(
-    val seasonId: Int,
-    val tvShowId: Int,
+    val seasonId: Long,
+    val tvShowId: Long,
     val name: String,
     val overview: String,
     val seasonNumber: Int,
